+++ packages/resource-time-grid/src/View.svelte
@@ -4,22 +4,32 @@
 	import {Header, Body, Day} from '@event-calendar/time-grid';
 
 	let state = getContext('state');
-	let {_viewDates, _intlDayHeader, theme} = state;
+	let {datesAboveResources, _viewDates, _intlDayHeader, theme} = state;
 
 	let viewState = new State(state);
 	setContext('view-state', viewState);
 
 	let {_viewResources} = viewState;
+
+	let loops, titles;
+	$: {
+		loops = [];
+		titles = [];
+		for (let i of $datesAboveResources ? [1, 0] : [0, 1]) {
+			loops.push(i ? $_viewDates : $_viewResources);
+			titles.push(i ? date => $_intlDayHeader.format(date) : resource => resource.title);
+		}
+	}
 </script>
 
 <Header>
-	{#each $_viewResources as resource}
+	{#each loops[0] as item0}
 		<div class="{$theme.resource}">
-			<div class="{$theme.day}">{resource.title}</div>
-			{#if $_viewDates.length > 1}
+			<div class="{$theme.day}">{titles[0](item0)}</div>
+			{#if loops[1].length > 1}
 				<div class="{$theme.days}">
-					{#each $_viewDates as date}
-						<div class="{$theme.day}">{$_intlDayHeader.format(date)}</div>
+					{#each loops[1] as item1}
+						<div class="{$theme.day}">{titles[1](item1)}</div>
 					{/each}
 				</div>
 			{/if}
@@ -27,10 +37,13 @@
 	{/each}
 </Header>
 <Body>
-	{#each $_viewResources as resource}
+	{#each loops[0] as item0}
 		<div class="{$theme.resource}">
-			{#each $_viewDates as date}
-				<Day {date} {resource}/>
+			{#each loops[1] as item1}
+				<Day
+						date={$datesAboveResources ? item0 : item1}
+						resource={$datesAboveResources ? item1 : item0}
+				/>
 			{/each}
 		</div>
 	{/each}